+++ core/src/commonMain/kotlin/com/github/jcornaz/kwik/generator/Sizes.kt
@@ -1,4 +1,6 @@
 package com.github.jcornaz.kwik.generator
 
-internal const val DEFAULT_MIN_SIZE = 0
-internal const val DEFAULT_MAX_SIZE = 200
+/**
+ * Default max size for string and collection-like generators.
+ */
+const val KWIK_DEFAULT_MAX_SIZE = 200