+++ core/src/commonMain/kotlin/com/github/jcornaz/kwik/generator/Strings.kt
@@ -1,6 +1,7 @@
 package com.github.jcornaz.kwik.generator
 
 import com.github.jcornaz.kwik.Generator
+import com.github.jcornaz.kwik.filterNot
 import com.github.jcornaz.kwik.withSamples
 
 @Suppress("MagicNumber", "TopLevelPropertyNaming")
@@ -13,8 +14,8 @@ private val PRINTABLE_CHARACTERS = (32..127).map { it.toChar() }.toSet()
  * @param exclude Characters to exclude from generated strings
  */
 fun Generator.Companion.strings(
-    minLength: Int = DEFAULT_MIN_SIZE,
-    maxLength: Int = maxOf(minLength, DEFAULT_MAX_SIZE),
+    minLength: Int = 0,
+    maxLength: Int = maxOf(minLength, KWIK_DEFAULT_MAX_SIZE),
     charset: Set<Char> = PRINTABLE_CHARACTERS,
     exclude: Set<Char> = emptySet()
 ): Generator<String> {
@@ -37,3 +38,27 @@ fun Generator.Companion.strings(
 
     return generator.withSamples(samples)
 }
+
+/**
+ * Returns a generator of String of length between 1 and [maxLength] (inclusive)
+ *
+ * @param charset Set of character to be used in generated strings
+ * @param exclude Characters to exclude from generated strings
+ */
+fun Generator.Companion.nonEmptyStrings(
+    maxLength: Int = KWIK_DEFAULT_MAX_SIZE,
+    charset: Set<Char> = PRINTABLE_CHARACTERS,
+    exclude: Set<Char> = emptySet()
+): Generator<String> = strings(1, maxLength, charset, exclude)
+
+/**
+ * Returns a generator of non-blank String of length between 1 and [maxLength] (inclusive)
+ *
+ * @param charset Set of character to be used in generated strings
+ * @param exclude Characters to exclude from generated strings
+ */
+fun Generator.Companion.nonBlankStrings(
+    maxLength: Int = KWIK_DEFAULT_MAX_SIZE,
+    charset: Set<Char> = PRINTABLE_CHARACTERS,
+    exclude: Set<Char> = emptySet()
+): Generator<String> = nonEmptyStrings(maxLength, charset, exclude).filterNot { it.isBlank() }